+++ 2. Exercises/0.Exam Preparation/03. Java OOP Exam - 12 Dec 2020/Backery/src/bakery/entities/drinks/Tea.java
@@ -3,7 +3,7 @@ package bakery.entities.drinks;
 public class Tea extends BaseDrink {
     private static final double TEA_PRICE = 2.50;
 
-    public Tea(String name, double portion, String brand) {
+    public Tea(String name, int portion, String brand) {
         super(name, portion, TEA_PRICE, brand);
     }
