+++ 2. Exercises/0.Exam Preparation/03. Java OOP Exam - 12 Dec 2020/Backery/src/bakery/entities/drinks/Water.java
@@ -5,7 +5,7 @@ import bakery.entities.drinks.BaseDrink;
 public class Water extends BaseDrink {
     private static final double WATER_PRICE = 1.50;
 
-    public Water(String name, double portion, String brand) {
+    public Water(String name, int portion, String brand) {
         super(name, portion, WATER_PRICE, brand);
     }
 }
